+++ src/components/index.js
@@ -8,3 +8,4 @@ export Table from './Table';
 export LoadingSpinner from './LoadingSpinner';
 export Searchbar from './Searchbar';
 export Modal from './Modal';
+export CheckBox from './CheckBox';